RAGSDK-master/requirements.txt
openpyxl==3.1.2
xlrd~=2.0.1
loguru==0.7.3
transformers~=4.41.2
sentence_transformers~=5.0.0
accelerate==0.30.1
torch==2.1.0
torchvision==0.16.0
cn_clip==1.5.1
numpy==1.26.4
urllib3~=2.4.0
pydantic==2.11.7
sqlalchemy~=2.0.23
python-docx==1.1.2
pillow==10.4.0
pymilvus~=2.5.0
PyMuPDF==1.21.0
paddleocr==2.9.1
paddlepaddle==2.6.2
opencv-python~=4.6.0.66
scikit-learn==1.6.1
umap-learn~=0.5.6
pyOpenSSL==25.3.0
cryptography==46.0.5
tqdm==4.66.5
jieba==0.42.1
python-pptx==0.6.23
langchain==0.3.27
gptcache==0.1.43
cachetools==5.3.3
pandas==2.2.2
langchain_community==0.3.27
opengauss-sqlalchemy==2.3.1
psycopg2==2.9.9
networkx==3.2.1
ragas==0.2.15
json-repair==0.44.1
einops~=0.8.0
unstructured==0.18.11
Markdown==3.9